cmscal: Get rid of buildid flag
In retrospect, this wasn't a good idea. The value was used to generate
IDs for calendar events, and I wanted those IDs to change any time the
server build did. But now that I think about it, a new build might
change the calendar object, but as long as the data in an event has a
stable fingerprint, its ID can remain.