bazel: Fix build stamping
This fixes a bug in commit 98f692e where a build wouldn't be re-run when code
didn't change but the Git HEAD did. Problem: we were writing a volatile tag
instead of a stable one (one that begins with STABLE_). The contract for build
stamping is that stable tags should change "rarely," and Bazel will rebuild
stamped targets whenever there's a change, which is exactly the behavior we
were looking for.